crewai_backend/requirements.txt
Flask>=2.2.3
flask-cors
python-dotenv
crewai
pyyaml
openai
tiktoken
requests